Tell two integers apart that a double cannot (#269) - Header and body were both converted to double before comparing, so   9007199254740992 and 9007199254740993 were one number: a header could   carry a value the server never read out of the body, which is the split   the mirroring exists to prevent - Integers now compare as integers, with the wider comparison kept for   the case it was for — a body carrying 42.0 and a header carrying 42 - And read with the 64-bit accessor throughout: the 32-bit one truncates   the very values this is about

+++ src/transport/streamable_http_client_session.cc
@@ -69,12 +69,12 @@ void StreamableHttpClientSession::decorate(
       continue;
     }
     if (value.isInteger() &&
-        !protocol::modern::isExactlyCarryableInteger(value.getInt())) {
+        !protocol::modern::isExactlyCarryableInteger(value.getInt64())) {
       GOPHER_LOG_WARN(
           "not mirroring {} into a header: {} cannot be carried exactly, and "
           "a value the two ends round differently is a value they disagree "
           "about",
-          param.headerName(), value.getInt());
+          param.headerName(), value.getInt64());
       continue;
     }
     std::string text;
